ASoC: rockchip: i2s-tdm: rename + fix frame-width
Previously this was called bclk-fs in the downstream driver, which
is not very descriptive. It also did not actually check for its
maximum upper bound, and simply silently discarded any values it
didn't like instead of loudly complaining.

Fix all of this. If something isn't within the range of allowed
values, return from the probe function with -EINVAL as one should.

Signed-off-by: Nicolas Frattaroli <frattaroli.nicolas@gmail.com>