ASoC: rockchip: i2s-tdm: rework trcm sync
First off, this renames clk-trcm to trcm-sync, and makes it a
required property. The rationale for this is that it's better
to be explicit as to which clock one wishes to synchronise to
than to implicitly get assigned some value.

This also cleans up the handling of how the property is read in
the probe function, not relying on the 0-ness of the read value
as much and instead going for a more obvious switch case approach
for branching depending on the value.

Instead of assigning magic values (0, 1, 2) to the property, we
provide a header with obvious names for what those things are.

This also updates the Quartz64 Model A device tree to using the
new property.

Signed-off-by: Nicolas Frattaroli <frattaroli.nicolas@gmail.com>